refactor(hub): split message and wake tests out of hub_test.go
Moves the conversation message/timeline/streaming tests to
message_handlers_test.go and the initial-plan/wake tests to
wake_test.go, mirroring the production file split. Code moved
verbatim; hub_test.go drops from 1,954 to 1,201 LOC (phase-2
ceiling is 1,500).
